+++ pos_order_cancel/models/models.py
@@ -1,19 +1,24 @@
 # -*- coding: utf-8 -*-
 from odoo import fields, models, api, _
 from odoo.exceptions import UserError
+from functools import partial
+from datetime import datetime
+from odoo.tools import DEFAULT_SERVER_DATETIME_FORMAT as DF
 
 
 class PosCancelledReason(models.Model):
     _name = "pos.cancelled_reason"
 
-    number = fields.Integer(string="Number")
+    sequence = fields.Integer(string="Sequence")
     name = fields.Char(string="Reason")
 
 
 class PosOrder(models.Model):
     _inherit = "pos.order"
-    cancellation_reason = fields.Text(string='Order Cancellation Reason')
+
+    cancellation_reason = fields.Text(string="Order Cancellation Reason")
     is_cancelled = fields.Boolean("Stage Cancelled", default=False)
+    canceled_lines = fields.One2many("pos.order.line.canceled", "order_id", string="Canceled Lines")
     computed_state = fields.Selection(
         [('draft', 'New'), ('cancel', 'Cancelled'), ('paid', 'Paid'), ('done', 'Posted'), ('invoiced', 'Invoiced')],
         'Status', compute='_compute_state')
@@ -28,150 +33,23 @@ class PosOrder(models.Model):
                 pos_order.computed_state = pos_order.state
 
     @api.model
+    def _order_fields(self, ui_order):
+        order = super(PosOrder, self)._order_fields(ui_order)
+        process_canceled_line = partial(self.env['pos.order.line.canceled']._order_cancel_line_fields)
+        order['canceled_lines'] = [process_canceled_line(l) for l in ui_order['canceled_lines']] if ui_order['canceled_lines'] else False
+        return order
+
+    @api.model
     def _process_order(self, pos_order):
         order = super(PosOrder, self)._process_order(pos_order)
         if 'is_cancelled' in pos_order and pos_order['is_cancelled'] is True:
-            order.cancellation_reason = str(pos_order['CancellationReason'].strip(" \t\n"))
+            order.cancellation_reason = str(pos_order['reason'].strip(" \t\n"))
             order.is_cancelled = True
         return order
 
     def _create_account_move_line(self, session=None, move=None):
-        # Tricky, via the workflow, we only have one id in the ids variable
-        """Create a account move line of order grouped by products or not."""
-        IrProperty = self.env['ir.property']
-        ResPartner = self.env['res.partner']
-
-        if session and not all(session.id == order.session_id.id for order in self):
-            raise UserError(_('Selected orders do not have the same session!'))
-
-        grouped_data = {}
-        have_to_group_by = session and session.config_id.group_by or False
-        rounding_method = session and session.config_id.company_id.tax_calculation_rounding_method
-
-        for order in self.filtered(lambda o: (not o.account_move or order.state == 'paid') and o.is_cancelled is False):
-            current_company = order.sale_journal.company_id
-            account_def = IrProperty.get(
-                'property_account_receivable_id', 'res.partner')
-            order_account = order.partner_id.property_account_receivable_id.id or account_def and account_def.id
-            partner_id = ResPartner._find_accounting_partner(order.partner_id).id or False
-            if move is None:
-                # Create an entry for the sale
-                journal_id = self.env['ir.config_parameter'].sudo().get_param(
-                    'pos.closing.journal_id_%s' % current_company.id, default=order.sale_journal.id)
-                move = self._create_account_move(
-                    order.session_id.start_at, order.name, int(journal_id), order.company_id.id)
-
-            def insert_data(data_type, values):
-                # if have_to_group_by:
-                values.update({
-                    'partner_id': partner_id,
-                    'move_id': move.id,
-                })
-
-                if data_type == 'product':
-                    key = ('product', values['partner_id'], (values['product_id'], tuple(values['tax_ids'][0][2]), values['name']), values['analytic_account_id'], values['debit'] > 0)
-                elif data_type == 'tax':
-                    key = ('tax', values['partner_id'], values['tax_line_id'], values['debit'] > 0)
-                elif data_type == 'counter_part':
-                    key = ('counter_part', values['partner_id'], values['account_id'], values['debit'] > 0)
-                else:
-                    return
-
-                grouped_data.setdefault(key, [])
-
-                if have_to_group_by:
-                    if not grouped_data[key]:
-                        grouped_data[key].append(values)
-                    else:
-                        current_value = grouped_data[key][0]
-                        current_value['quantity'] = current_value.get('quantity', 0.0) + values.get('quantity', 0.0)
-                        current_value['credit'] = current_value.get('credit', 0.0) + values.get('credit', 0.0)
-                        current_value['debit'] = current_value.get('debit', 0.0) + values.get('debit', 0.0)
-                else:
-                    grouped_data[key].append(values)
-
-            # because of the weird way the pos order is written, we need to make sure there is at least one line,
-            # because just after the 'for' loop there are references to 'line' and 'income_account' variables (that
-            # are set inside the for loop)
-            # TOFIX: a deep refactoring of this method (and class!) is needed
-            # in order to get rid of this stupid hack
-            assert order.lines, _('The POS order must have lines when calling this method')
-            # Create an move for each order line
-            cur = order.pricelist_id.currency_id
-            for line in order.lines:
-                amount = line.price_subtotal
-
-                # Search for the income account
-                if line.product_id.property_account_income_id.id:
-                    income_account = line.product_id.property_account_income_id.id
-                elif line.product_id.categ_id.property_account_income_categ_id.id:
-                    income_account = line.product_id.categ_id.property_account_income_categ_id.id
-                else:
-                    raise UserError(_('Please define income '
-                                      'account for this product: "%s" (id:%d).')
-                                    % (line.product_id.name, line.product_id.id))
-
-                name = line.product_id.name
-                if line.notice:
-                    # add discount reason in move
-                    name = name + ' (' + line.notice + ')'
-
-                # Create a move for the line for the order line
-                insert_data('product', {
-                    'name': name,
-                    'quantity': line.qty,
-                    'product_id': line.product_id.id,
-                    'account_id': income_account,
-                    'analytic_account_id': self._prepare_analytic_account(line),
-                    'credit': ((amount > 0) and amount) or 0.0,
-                    'debit': ((amount < 0) and -amount) or 0.0,
-                    'tax_ids': [(6, 0, line.tax_ids_after_fiscal_position.ids)],
-                    'partner_id': partner_id
-                })
-
-                # Create the tax lines
-                taxes = line.tax_ids_after_fiscal_position.filtered(lambda t: t.company_id.id == current_company.id)
-                if not taxes:
-                    continue
-                for tax in taxes.compute_all(line.price_unit * (100.0 - line.discount) / 100.0, cur, line.qty)['taxes']:
-                    insert_data('tax', {
-                        'name': _('Tax') + ' ' + tax['name'],
-                        'product_id': line.product_id.id,
-                        'quantity': line.qty,
-                        'account_id': tax['account_id'] or income_account,
-                        'credit': ((tax['amount'] > 0) and tax['amount']) or 0.0,
-                        'debit': ((tax['amount'] < 0) and -tax['amount']) or 0.0,
-                        'tax_line_id': tax['id'],
-                        'partner_id': partner_id
-                    })
-
-            # round tax lines per order
-            if rounding_method == 'round_globally':
-                for group_key, group_value in grouped_data.iteritems():
-                    if group_key[0] == 'tax':
-                        for line in group_value:
-                            line['credit'] = cur.round(line['credit'])
-                            line['debit'] = cur.round(line['debit'])
-
-            # counterpart
-            insert_data('counter_part', {
-                'name': _("Trade Receivables"),  # order.name,
-                'account_id': order_account,
-                'credit': ((order.amount_total < 0) and -order.amount_total) or 0.0,
-                'debit': ((order.amount_total > 0) and order.amount_total) or 0.0,
-                'partner_id': partner_id
-            })
-
-            order.write({'state': 'done', 'account_move': move.id})
-
-        all_lines = []
-        for group_key, group_data in grouped_data.iteritems():
-            for value in group_data:
-                all_lines.append((0, 0, value),)
-        if move:  # In case no order was changed
-            move.sudo().write({'line_ids': all_lines})
-            move.sudo().post()
-        return True
+        canceled_order = self.filtered(lambda o: (not o.account_move or o.state == 'paid') and o.is_cancelled is True)
+        return super(PosOrder, self-canceled_order)._create_account_move_line(session, move)
 
     @api.multi
     def action_pos_order_paid(self):
@@ -188,3 +66,32 @@ class PosOrder(models.Model):
             return True
         else:
             return super(PosOrder, self).action_pos_order_done()
+
+
+class PosOrderLineCanceled(models.Model):
+    _name = "pos.order.line.canceled"
+    _rec_name = "product_id"
+
+    def _order_cancel_line_fields(self, line):
+        if line and 'tax_ids' not in line[2]:
+            product = self.env['product.product'].browse(line[2]['product_id'])
+            line[2]['tax_ids'] = [(6, 0, [x.id for x in product.taxes_id])]
+        return line
+
+    product_id = fields.Many2one('product.product', string='Product', domain=[('sale_ok', '=', True)], required=True, change_default=True)
+    discount = fields.Float(string='Discount (%)', digits=0, default=0.0)
+    price_unit = fields.Float(string='Unit Price', digits=0)
+    user_id = fields.Many2one(comodel_name='res.users', string='Salesman', help="Person who removed order line", default=lambda self: self.env.uid)
+    qty = fields.Float('Quantity', default=1)
+    reason = fields.Text(string="The Reason of Line Canceled")
+    order_id = fields.Many2one('pos.order', string='Order Ref', ondelete='cascade')
+    pack_lot_ids = fields.One2many('pos.pack.operation.lot', 'pos_order_line_id', string='Lot/serial Number')
+    tax_ids = fields.Many2many('account.tax', string='Taxes', readonly=True)
+    canceled_date = fields.Datetime(string='Canceled Date', readonly=True)
+
+    @api.model
+    def create(self, values):
+        if values.get('canceled_date'):
+            canceled_date = datetime.strptime(values.get('canceled_date'), "%d/%m/%Y %H:%M:%S")
+            values['canceled_date'] = canceled_date.strftime(DF)
+        return super(PosOrderLineCanceled, self).create(values)
