+++ pos_order_cancel/static/src/xml/cancel_order.xml
@@ -1,20 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates id="template" xml:space="preserve">
 
-    <t t-extend="OrderChangeReceipt">
-        <t t-jquery='t[t-foreach="changes.cancelled"]' t-operation="after">
-            <t t-if="changes.reason">
-                <br />
-                <line>
-                    <left>REASON</left>
-                    <right>...</right>
-                </line>
-                <div><span font='b' bold='off' line-ratio='0.4' indent='1' width='30'>--- <t t-esc="changes.reason" /></span></div>
-                <line></line>
-            </t>
-        </t>
-    </t>
-
     <t t-name="ConfirmCancellationPopupWidget">
         <div class="modal-dialog">
             <div class="popup popup-confirm-cancellation">
