+++ pos_order_cancel/views/template.xml
@@ -3,7 +3,8 @@
     <data>
         <template id="assets_backend" name="pos_category_multi_assets" inherit_id="point_of_sale.assets">
             <xpath expr="//link[last()]" position="after">
-                <script type="text/javascript" src="/pos_order_cancel/static/src/js/cancel_order.js"></script>
+                <script type="text/javascript" src="/pos_order_cancel/static/src/js/models.js"></script>
+                <script type="text/javascript" src="/pos_order_cancel/static/src/js/widgets.js"></script>
                 <link rel="stylesheet" href="/pos_order_cancel/static/src/css/pos.css"/>
             </xpath>
         </template>
