+++ pos_order_cancel/views/views.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0"?>
-<openerp>
+<odoo>
     <record id="pos_cancelled_reason_form_view" model="ir.ui.view">
         <field name="name">pos.cancelled_reason.form</field>
         <field name="model">pos.cancelled_reason</field>
@@ -7,7 +7,6 @@
             <form string="POS Product Cancellation Reason">
                 <sheet>
                     <group>
-                        <field name="number"/>
                         <field name="name"/>
                     </group>
                 </sheet>
@@ -20,7 +19,7 @@
         <field name="model">pos.cancelled_reason</field>
         <field name="arch" type="xml">
             <tree string="Pos Product Cancellation Reason">
-                <field name="number"/>
+                <field name="sequence" widget="handle"/>
                 <field name="name"/>
             </tree>
         </field>
@@ -65,6 +64,33 @@
             <xpath expr="//field[@name='state']" position="attributes">
                 <attribute name="invisible">True</attribute>
             </xpath>
+
+            <xpath expr="//notebook" position="inside">
+                <page string="Canceled Products">
+                    <field name="canceled_lines" colspan="4" nolabel="1">
+                        <tree string="Canceled Order lines" editable="bottom">
+                            <field name="product_id"/>
+                            <field name="price_unit" widget="monetary"/>
+                            <field name="discount" widget="monetary"/>
+                            <field name="reason" string="Reason"/>
+                            <field name="canceled_date" string="Canceled Datetime"/>
+                            <field name="tax_ids" invisible="1"/>
+                        </tree>
+                        <form string="Order lines">
+                            <group col="4">
+                                <field name="product_id"/>
+                                <field name="price_unit" widget="monetary"/>
+                                <field name="discount" widget="monetary"/>
+                                <field name="reason" string="Reason"/>
+                                <field name="canceled_date" string="Canceled Datetime"/>
+                                <field name="tax_ids"/>
+                                <field name="user_id"/>
+                            </group>
+                        </form>
+                    </field>
+                    <div class="oe_clear"/>
+                </page>
+            </xpath>
         </field>
     </record>
 
@@ -101,5 +127,4 @@
             </xpath>
         </field>
     </record>
-
-</openerp>
+</odoo>
